fix selenium version numbers
main issue is there are large compatibility issues and sketchy sources for getting the chromium app  and the driver.

Docker now uses v90 driver with v90 bin. this is because the docker image is expecting deb instance with its python:3.8

also holy cow this was a nightmare to find a decent guide for

sh. startup uses v92.